error fixed

+++ src/main/java/com/b2b/b2b/modules/crm/deal/payloads/DealResponseDTO.java
@@ -3,12 +3,13 @@ package com.b2b.b2b.modules.crm.deal.payloads;
 import com.b2b.b2b.modules.crm.company.payloads.CompanyResponseDTO;
 import com.b2b.b2b.modules.crm.deal.entity.DealStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record DealResponseDTO(
         Integer id,
         String dealName,
-        Double dealAmount,
+        BigDecimal dealAmount,
         DealStatus dealStatus,
         String stageName,
         Integer stageOrder,
